requirements.txt
torch>=2.0.0
numpy>=1.23.0
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
tensorboard>=2.14.0